+++ tests/integration/generic_structs_phase5_core_helper.c
@@ -8,8 +8,8 @@
 #include "ast_operations.h"
 
 // Global test counters
-int tests_run = 0;
-int tests_passed = 0;
+size_t tests_run = 0;
+size_t tests_passed = 0;
 
 // ============================================================================
 // COMPILATION PIPELINE IMPLEMENTATION
