+++ tests/integration/generic_structs_test_common.h
@@ -34,8 +34,8 @@ extern "C" {
 // TEST FRAMEWORK MACROS
 // =============================================================================
 
-extern int tests_run;
-extern int tests_passed;
+extern size_t tests_run;
+extern size_t tests_passed;
 
 #define TEST_ASSERT(condition, message) do { \
     tests_run++; \
