+++ tests/integration/test_concurrency_tiers_common.c
@@ -12,8 +12,8 @@
 #include "ast_operations.h"
 
 // Global test counters
-int tests_run = 0;
-int tests_passed = 0;
+size_t tests_run = 0;
+size_t tests_passed = 0;
 
 // ============================================================================
 // Helper Functions Implementation
